move pin & schema out of stream folder
