+++ packages/cli/src/commands/pin/rm.ts
@@ -1,9 +1,9 @@
 import { StreamID } from '@ceramicnetwork/streamid'
 
-import { Command } from '../../../command'
-import type { CommandFlags } from '../../../command'
+import { Command } from '../../command'
+import type { CommandFlags } from '../../command'
 
-export default class Add extends Command<
+export default class Remove extends Command<
   CommandFlags,
   {
     streamId: string
